Fixed an illegal type comparison, added testing for filterTodosByStatus
Previously todoDatabase had a .equals comparison between x.status and target status... where both are booleans. Changed to == for proper comparison.\

Created two tests for filtering todos by status. One ensures that the expected number of todos with "complete" status are pulled, and the other ensures the same thing but for "incomplete" statuses.

All previous functionality seems to be intact and all tests are passing.