Implement filtering todos by category and fixed copy-paste error
Allows users to enter in a category on the website in the category box, and restrict todo search results to said category.

Added testing for filterTodosByCategory where the filtered todos are checked to make sure they are of the expected length.

Added testing for TodoControllerSpec as well.

All tests are passing and coverage remains at 100% for all but the server.

Closes #7